Add healthchecks to compose setup (#55)
- Moved `dev` entrypoint to `tini` (via `init: true`) and change command
to `sleep inf`.
- This means that `SIGTERM` triggers a shutdown rather than needing to
wait for `SIGKILL` 30s later.
- Added a small CLI for to support rpc healthchecks.
- Added health checks for the `rpc`, `grafana`, `jaeger`, and
`victoria-metrics` containers.
- Add `--wait` to `dev-up` recipe to confirm health checks are passing.
- Require `jaeger` and `victoria-metrics` to be healthy before starting
`rpc`.
- Define start period as 30s to enable use of a much reduced start
interval of 1s.
- By default health is polled every 30s, which the initial healthy
status takes unnecessarily long to come in. By defining a start period,
a more aggressive poll interval can be set until the first healthy is
reported and so provide a faster start.

The `dev` container does not get a health check as it's not really doing
anything until a user exec's into it.

Fixes #37.