Implement tag action (#85)
Implement the creation of tags in the tag action such that each commit
to the default branch is tagged. This is intended for providing a human
readable version string and is not compatible with semantic versioning.
However, it can be appropriate for versioning microservices when the
service APIs are semantically versioned.

The action is intended to operate in one of 2 modes when creating tags
depending on whether the `rollover` input is `true` or `false`:

- When `rollover=false`, forever increment the last segment of the
version, e.g. `0.0.1 -> 0.0.2`.
- When `rollover=true`, rollover the when the increment would reach a
limit, e.g. `0.0.99 -> 0.1.0`, `0.99.99 -> `1.0.0`.

The point at which rollover occurs is defined by the limit input.

The rollover logic can be intercepted or performed manually by setting
the `minimum` input to a higher version than the the automatically
calculated version. The `minimum` input can also be used to implement
alternative versioning schemes, e.g. a `minimum` of `<year>.<month>.0`
would provide a date-based versioning scheme.