Workaround zizmor dockerfile using wolfi-base (#221)
The zizmor container changed to use wolfi-base as it's base image,
installing zizmor via apk. While this removed bash (see #219), the CI
entrypoint script was not too hard to adapt.

It did however affect the dev-env. The dev-env uses Debian 12, which
offers a relatively stable set of dependencies. Wolfi (and it's
maintainers at Chainguard) meanwhile prioritises security over
stability, and so will regularly update OS dependencies. This includes
glibc.

The problem is that zizmor is not statically linked by whatever builds
the zizmor apk, and so depends on glibc 2.39. Meanwhile, debian 12 comes
with 2.36. Glibc is often backwards compatible, but never forwards
compatible, and so the rust runtime aborts when executing zizmor. While
we could wait for debian 13 it just moves the problem down the road
until wolfi updates again.

To workaround the issue, I've swapped the install of zizmor to pip. This
opens up another can of worms in how debian 12 works with pip - a
cursory glance suggests it mostly doesn't work and python3-pip is likely
useless now. but with a little brute force, it can be used to install
zizmor. pipx seems to be a better dependency manager for this, but I
didn't investigate much further, especially as zizmor doesn't appear to
have any python dependencies.

This does introduce a potential drift between the dev-env and CI. this
can possibly be resolved by doing the opposite of what golangci-lint
does, i.e. copy the version from the CI dockerfile and pass it as a
build arg. But that can come later - on the off chance better ideas
materialise such as unifying the install method for both workflows.