Align gRPC dashboard's internal errors with ConnectRPC dashboard (#230)
Adds Unimplemented and Data Loss to the gRPC status codes considered to
be an internal error in the gRPC dashboard. This aligns with the
ConnectRPC dashboard, as well as the behaviour in the otelconnect
interceptor which marks the span status as error for 6 status codes.

Also selects specific tags for images in the development environment,
resolving a handful of TODOs.

Fixes #196.