Export RPC logs to VictoriaLogs via OTLP (#252)
This change configures an OpenTelemetry logger provider to export logs
in batches to VictoriaLogs via OTLP. VictoriaLogs can then be used by
Grafana to view metrics, traces and (now) logs.

To achieve this, an additional slog handler has been configured that
handles logs in parallel to the existing JSON logger. This means that
the JSON logs continue to be output to STDOUT for debugging when needed,
but the logs are also passed to VictoriaLogs. The format of the logs is
different as OpenTelemetry defines it's own format. But the output is
just as useful.

This does not address capturing logs for all services in the development
environment. As noted in #117, doing so was excessively complicated.
While it would be nice to have postgres, etc. logs in Grafana too, RPC
logs are the most interesting to us in reality. Additionally, just
having RPC logs is consistent with just having RPC metrics and traces in
Grafana too.

Fixes #117.