Implement baggage propagation (#278)
Add support for baggage propagation to enable functional tests to pass
the test scenario name to the server. Also copy baggage into log and
span attributes using the OpenTelemetry SDK.

- Store the test scenario name in baggage, replacing spaces with
underscores.
- Configure baggage propagation in the test ConnectRPC client.
- Configure baggage and trace propagation in the ConnectRPC server.
- Configure a baggage span processor to add baggage as span attributes.
- Configure a baggage log processor to add baggage as log attributes.

Due to baggage log processing being part of the OpenTelemetry ecosystem,
baggage is not included on logs output to stdout. Instead, baggage is
only added when exporting logs via OTLP. A brief investigation was
conducted into whether logs could also be exported to stdout and gain
baggage that way, but the output format was excessively verbose and
significantly reduce debuggability.

At present, all baggage members are added to logs and spans. This can be
reduced to a statis or dynamic allowlist by re-implementing the filter
added in `wire_providers.go`.

Fixes #218.