Switch to gofrs as UUID library (#342)
Move from `github.com/google/uuid` to `github.com/gofrs/uuid/v5` for the
following reasons:

- gofrs is more actively maintained and contributed to.
- With recent layoffs, it's not clear that a maintainer exists for the
google library.
- gofrs implements generation via a struct meaning we no longer need the
very thin uuidgen driver.

`jet` has a default dependency on the google package, which makes it a
little tricky to fully remove. Happily, this is customisable, and will
be fully transitioned in #353. Until then, there is a small set of
translation functions in the datastore adapter to convert one to the
other. The data structure layout is identical in both (for now) and unit
tests have been added to ensure that doesn't unexpectedly change.