Fix k8smount koanf provider to actually work with volume mounts (#352)
Fix the k8smount koanf provider to actually work with volume mounts of
secrets and configmaps.

Until now, the correctness of this provider has been limited to unit
tests and had not been tested against actual Kubernetes. However, with
the progress in #323, testing found that it chocked on the symlink
structure created by Kubernetes, particularly a symlink to a directory.

This change now moves towards correctly reading configuration for data
in a flat directory structure. Despite existing assertions that data
cannot be in subdirectories, it seems to be possible by setting a path
in the items array of a secret or configmap volume on a pod. Support for
this will be deferred to a future PR.