Use customised jet generation to modify UUID type in database models (#353)
This introduces a CLI for generated code based on the database schema.
This is a thin wrapper around `jet` which was previously invoked via CLI
directly. This wrapping allows some small customisations to be
introduced, removes a tool managed by `bingo`, and builds upon the
existing driver for PostgreSQL connections. While this change does not
add support for tracing or metrics, they could be added if they ever
become valuable.

Full list of changes are:

- Override the `UUID` type used by `jet` generation to
`github.com/gofrs/uuid/v5`.
- Reduce the depth of the package(s) `jet` outputs code to.
- Remove obsolete UUID type conversions.
- Change the type of `interna;/drivers/pgsql.Config.Port` int for better
type safety.
- Remove `jet` from bingo's configuration.

Fixes #350.