Set GOTOOLCHAIN=local in container builds (#380)
As of Go 1.21, Go will automatically use a newer compiler version if the
build requires it as documented in [Go
Toolchains](https://go.dev/doc/toolchain). For example, running `go
install` with Go 1.21 against a module that require Go 1.23, Go 1.23
will be used instead.

This feature does not seem unreasonable for tool installs which can move
ahead of the project's Go version. However, it is problematic for
container builds which (in this project) are supposed to be hermetic.
Having an unexpected variable means it is difficult to check what
version of the compiler will be used.

To disable this feature, we can set `GOTOOLCHAIN=local`, which replaces
the default `GOTOOLCHAIN=local+auto`.