Run `just gen-just` before `just checks` in regenerate workflow (#385)
When bingo-managed tools are updated, `just gen-just` sets the new
version into `.tools.just`. However, because `just` is still running, it
does not update its variables to point to the new tools and subsequent
recipes fails with the wrong version.

To fix this, we can run the `gen-just` recipe first. This is performed
in a separate process to disable any re-use of outdated variables.

Happily, this change also fixes the outdated `.bingo/*.sum` files which
Renovate doesn't seem to know how to update.