Expose monotonic time in Clock interface (#409)
As discussed in [Monotonic and Wall Clock Time in the Go time
package][1], `time.Now().UTC()` strips the monotonic component of
`time.Time` making it unsuitable for calculating the duration between 2
points in time. To correct this the `repositories.Clock.Now` method has
been renamed to `UTCNow` and a new `LocalNow` has been added to return
`time.Time` with a monotonic component.

This does not impact the application as the `Since` and `Until` methods
of `repositories.Clock` were unused, but it's a useful improvement
regardless.

Fixes #408.

[1]: https://victoriametrics.com/blog/go-time-monotonic-wall-clock/