Remove use of wire (#449)
[wire] has been archived, likely due to layoffs at Google, and it has
been largely unmaintained for some years. As of Go 1.25 it is broken and
needs an update to `golang.org/x/tools` to become functional again.
Obviously it's archival makes that much harder.

Rather than fork and fix it, this change simply removes it and replaces
what it was doing with a series of setup functions. This works our as
roughly the same amount of work as maintaining wire was previously given
the various workaround needed to use wire.

[wire]: https://github.com/google/wire