Replace `slogt` dependency (#453)
The `slogt` dependency was a useful way of getting `slog.Logger`
instances to output to the same writer as `testing.T.Log`, which was a
moderately involved process. However, with the release of Go 1.25, we
can use `testing.T.Output` to obtain that writer and pass it directly to
a new `slog.Handler`. There is so little code required to support this
that it seems pointless to keep the dependency and so its been replaced.

- Upgrade to Go 1.25.
- Replace `github.com/neilotoole/slogt` with `pkg/slogt`.
- Update imports.