Create Tilt environment with observability charts (#488)
Initialise the `Tiltfile` with observability tooling, namely Tempo,
VictoriaMetrics and VictoriaLogs. This is a slight divergence from the
current use of Jaeger, but it gives us access to TraceQL. In the future,
we may switch to VictoriaTraces for consistency.

The database and locally-developed services have been omitted at this
time. Development of those can be found in #323.

This also includes a custom manager in Renovate to enable automatic
updates to the charts. The correctness of this will be validated
post-merge.