Use underscores for splitting configmap fields (#509)
While working on dynamically generating a ConfigMap using Helm's
templating functionality, I realised that the period delimited keys in
ConfigMaps is somewhat unnecessary. For environment variables we
unconditionally split on underscores, and so it seems more consistent to
match that precedent and do the same in ConfigMap and Secret volume
mounts.

Full changes:

- Update sanity checks in README to include authentication.
- Change the delimiter for splitting volume mount fields from periods to
underscores.
- Fully document the config splitters package.
- Convert an empty string to an empty slice in the splitter package and
add unit tests to validate the behaviour.